more clear and readable code
